pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "nerdouts"
version = "0.1.0"
description = "A workout management system"
readme = "README.md"
requires-python = ">=3.7"
license = {text = "Creative Commons Attribution-Noncommercial-Share Alike license"}
authors = [
    {name = "Daniel Seisun"}
]
urls = {Homepage = "https://github.com/dseisun/nerdouts"}
dependencies = [
    "pyyaml>=4.2b1",
    "SQLAlchemy",
]

[project.optional-dependencies]
linux = [
    "dbus-python==1.2.4"
]
mysql = [
    "mysqlclient==1.4.6"
]

postgres = [
    "psycopg2==2.9.10"
]

[tool.setuptools]
packages = ["nerdouts"]

[tool.setuptools.package-data]
nerdouts = ["*"]

[project.scripts]
nerdouts = "nerdouts.run_workout:main"
